series_project/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.1
funcy==2.0
redis==5.0.4
sqlparse==0.5.0
